Рефакторинг репозиториев
Resolve https://github.com/stas12312/my-tickets-bot/issues/25
