+++ my_tickets_bot/src/services/repositories/__init__.py
@@ -1,0 +1,2 @@
+"""Репозитории для сущностей"""
+from .repo import Repo
