+++ my_tickets_bot/src/services/repositories/queries/__init__.py
@@ -1,0 +1,1 @@
+"""Запросы"""
