+++ my_tickets_bot/src/services/repositories/queries/user.py
@@ -1,0 +1,9 @@
+"""Запросы для пользователей"""
+
+SAVE_USER = """
+    INSERT 
+    INTO "user" (id, username) VALUES ($1, $2)
+    ON CONFLICT (id) DO UPDATE
+    SET id=$1, username=$2
+    RETURNING *
+    """
